// types/contribution.ts
export interface Contribution {
    module:               string;
    id:                   number;
    entity:               null;
    import_key:           null;
    array_options:        any[];
    array_languages:      null;
    contacts_ids:         null;
    linked_objects:       null;
    linkedObjectsIds:     null;
    oldref:               null;
    project:              Project;
    fk_project:           number;
    contact_id:           null;
    user:                 null;
    origin:               null;
    origin_id:            null;
    ref:                  string;
    ref_ext:              null;
    statut:               null;
    status:               number;
    country_id:           null;
    country_code:         null;
    state_id:             null;
    region_id:            null;
    barcode_type:         null;
    barcode_type_coder:   null;
    mode_reglement_id:    null;
    cond_reglement_id:    null;
    demand_reason_id:     null;
    transport_mode_id:    null;
    shipping_method_id:   null;
    shipping_method:      null;
    multicurrency_code:   null;
    multicurrency_tx:     null;
    model_pdf:            null;
    last_main_doc:        null;
    last_main_doc_url:    null;
    fk_bank:              null;
    fk_account:           null;
    note_public:          string;
    note_private:         string;
    total_ht:             null;
    total_tva:            null;
    total_localtax1:      null;
    total_localtax2:      null;
    total_ttc:            null;
    lines:                null;
    name:                 null;
    lastname:             null;
    firstname:            null;
    civility_id:          null;
    date_creation:        number;
    date_validation:      null;
    date_modification:    null;
    date_update:          null;
    date_cloture:         null;
    user_author:          null;
    user_creation:        null;
    user_creation_id:     null;
    user_valid:           null;
    user_validation:      null;
    user_validation_id:   null;
    user_closing_id:      null;
    user_modification:    null;
    user_modification_id: null;
    specimen:             number;
    labelStatus:          null;
    showphoto_on_popup:   null;
    nb:                   any[];
    output:               null;
    extraparams:          any[];
    label:                string;
    amount:               number;
    qty:                  number;
    fk_soc:               null;
    description:          string;
    tms:                  number;
    fk_user_creat:        number;
    fk_user_modif:        null;
    fk_community:         number;
    community:            Community;
}

interface Community {
    module:                        string;
    id:                            number;
    entity:                        null;
    import_key:                    string;
    array_options:                 any[];
    array_languages:               null;
    contacts_ids:                  null;
    linked_objects:                null;
    linkedObjectsIds:              null;
    oldref:                        null;
    project:                       Project;
    fk_project:                    string;
    contact_id:                    null;
    user:                          null;
    origin:                        null;
    origin_id:                     null;
    ref:                           string;
    company_name:                  string | null;
    customer_type:                 1 | 2 | 3;
    ref_ext:                       null;
    statut:                        null;
    status:                        number;
    country_id:                    null;
    country_code:                  string;
    state_id:                      null;
    region_id:                     null;
    barcode_type:                  null;
    barcode_type_coder:            null;
    mode_reglement_id:             null;
    cond_reglement_id:             null;
    demand_reason_id:              null;
    transport_mode_id:             null;
    shipping_method_id:            null;
    shipping_method:               null;
    multicurrency_code:            null;
    multicurrency_tx:              null;
    model_pdf:                     null;
    last_main_doc:                 string;
    last_main_doc_url:             null;
    fk_bank:                       null;
    fk_account:                    null;
    note_public:                   null;
    note_private:                  null;
    total_ht:                      null;
    total_tva:                     null;
    total_localtax1:               null;
    total_localtax2:               null;
    total_ttc:                     null;
    lines:                         null;
    name:                          null;
    lastname:                      string;
    firstname:                     string;
    civility_id:                   null;
    date_creation:                 number;
    date_validation:               null;
    date_modification:             null;
    date_update:                   null;
    date_cloture:                  null;
    user_author:                   null;
    user_creation:                 null;
    user_creation_id:              null;
    user_valid:                    null;
    user_validation:               null;
    user_validation_id:            null;
    user_closing_id:               null;
    user_modification:             null;
    user_modification_id:          null;
    specimen:                      number;
    labelStatus:                   null;
    showphoto_on_popup:            null;
    nb:                            any[];
    output:                        null;
    extraparams:                   any[];
    photo:                         string;
    display_name:                  "0" | "1";
    phone:                         string;
    email:                         string;
    address:                       string;
    rue:                           string;
    zip:                           string;
    town:                          string;
    fk_categories:                 string;
    date_carbone_killer:           string;
    lieu:                          string;
    profession:                    string;
    motCles:                       string;
    url:                           string;
    priorite_climatique:           string;
    bio:                           string;
    declic_climatique:             string;
    date_last_activity:            number;
    nombre_orders:                 null;
    fk_soc:                        null;
    description:                   null;
    tms:                           number;
    fk_user_author:                number;
    fk_user_modif:                 number;
    sexe:                          null;
    catched_carbon:                null;
    restored_biodiversity:         null;
    climatic_influence_citoyen:    null;
    climatic_influence_entreprise: null;
    photo_url:       string;
    photo_small_url: string;
    photo_mini_url:  string;
}

interface Project {
    module:                        null;
    id:                            string;
    entity:                        string;
    import_key:                    null;
    array_options:                 ArrayOptions;
    array_languages:               null;
    contacts_ids:                  null;
    linked_objects:                null;
    linkedObjectsIds:              null;
    oldref:                        null;
    fk_project:                    null;
    contact_id:                    null;
    user:                          null;
    origin:                        null;
    origin_id:                     null;
    ref:                           string;
    ref_ext:                       null;
    statut:                        string;
    status:                        string;
    country_id:                    null;
    country_code:                  null;
    state_id:                      null;
    region_id:                     null;
    barcode_type:                  null;
    barcode_type_coder:            null;
    mode_reglement_id:             null;
    cond_reglement_id:             null;
    demand_reason_id:              null;
    transport_mode_id:             null;
    shipping_method_id:            null;
    shipping_method:               null;
    multicurrency_code:            null;
    multicurrency_tx:              null;
    model_pdf:                     null;
    last_main_doc:                 null;
    last_main_doc_url:             null;
    fk_bank:                       null;
    fk_account:                    null;
    note_public:                   null;
    note_private:                  null;
    total_ht:                      null;
    total_tva:                     null;
    total_localtax1:               null;
    total_localtax2:               null;
    total_ttc:                     null;
    lines:                         null;
    name:                          null;
    lastname:                      null;
    firstname:                     null;
    civility_id:                   null;
    date_creation:                 null;
    date_validation:               null;
    date_modification:             null;
    date_update:                   null;
    date_cloture:                  null;
    user_author:                   null;
    user_creation:                 null;
    user_creation_id:              null;
    user_valid:                    null;
    user_validation:               null;
    user_validation_id:            null;
    user_closing_id:               null;
    user_modification:             null;
    user_modification_id:          string;
    specimen:                      number;
    labelStatus:                   null;
    showphoto_on_popup:            null;
    nb:                            any[];
    output:                        null;
    extraparams:                   any[];
    table_element_date:            null;
    description:                   string;
    title:                         string;
    dateo:                         null;
    date_start:                    number;
    datee:                         null;
    date_end:                      string;
    date_start_event:              string;
    date_end_event:                string;
    location:                      string;
    date_close:                    string;
    socid:                         null;
    thirdparty_name:               null;
    user_author_id:                string;
    fk_user_close:                 null;
    user_close_id:                 null;
    public:                        string;
    budget_amount:                 string;
    usage_opportunity:             number;
    usage_task:                    number;
    usage_bill_time:               number;
    usage_organize_event:          number;
    accept_conference_suggestions: number;
    accept_booth_suggestions:      number;
    price_registration:            null;
    price_booth:                   null;
    max_attendees:                 null;
    opp_status:                    string;
    opp_status_code:               string;
    opp_status_label:              string;
    fk_opp_status:                 null;
    opp_amount:                    string;
    opp_percent:                   string;
    opp_weighted_amount:           null;
    email_msgid:                   null;
    weekWorkLoad:                  null;
    weekWorkLoadPerTask:           null;
    datec:                         number;
    date_c:                        number;
    datem:                         number;
    date_m:                        number;
    photo:                         string;
    photo_url:                     string;
    photo_small_url:               string;
    photo_mini_url:                string;
    header_photo:                  null;
    header_photo_url:              null;
    profiles:                      any[];
}